refactor: locationDataManager을 DataManager에서 관리하도록 리팩터
