+++ WeatherForecast/WeatherForecast/Location/LocationDataManager.swift
@@ -35,7 +35,7 @@ extension LocationDataManager: CLLocationManagerDelegate {
             case .authorizedWhenInUse:
                 manager.startUpdatingLocation()
             case .restricted, .denied:
-                locationDelegate?.viewRequestLocationSettingAlert()
+                locationDelegate?.notifyDeniedAuthorization()
             case .notDetermined:
                 manager.requestWhenInUseAuthorization()
             default:
