+++ WeatherForecast/WeatherForecast/Location/LocationDataManagerDelegate.swift
@@ -10,5 +10,5 @@ import CoreLocation
 protocol LocationDataManagerDelegate: AnyObject {
     func location(_ manager: LocationDataManager, didLoadCoordinate location: CLLocationCoordinate2D)
     func loaction(_ manager: LocationDataManager, didCompletePlcamark location: CLPlacemark?)
-    func viewRequestLocationSettingAlert()
+    func notifyDeniedAuthorization()
 }
